Page 11h and float64s in config files

+++ opers/config.go
@@ -8,29 +8,29 @@ import (
 )
 
 type Module struct {
-	SFF8024Identifier                  int    `yaml:"SFF8024Identifier"`
-	CmisRevision                       int    `yaml:"CmisRevision"`
-	ModuleRevision                     int    `yaml:"ModuleRevision"`
-	MediaType                          int    `yaml:"MediaType"`
-	VendorName                         string `yaml:"VendorName"`
-	DateCode                           string `yaml:"DateCode"`
-	MaxPower                           int    `yaml:"MaxPower"`
-	LenghtSMF                          int    `yaml:"LenghtSMF"`
-	ModuleTempMax                      int    `yaml:"ModuleTempMax"`
-	ModuleTempMin                      int    `yaml:"ModuleTempMin"`
-	TempMonHighWarningThreshold        int    `yaml:"TempMonHighWarningThreshold"`
-	TempMonLowWarningThreshold         int    `yaml:"TempMonLowWarningThreshold"`
-	VccMonHighWarningThreshold         int    `yaml:"VccMonHighWarningThreshold"`
-	VccMonLowWarningThreshold          int    `yaml:"VccMonLowWarningThreshold"`
-	OpticalPowerTxHighWarningThreshold int    `yaml:"OpticalPowerTxHighWarningThreshold"`
-	OpticalPowerTxLowWarningThreshold  int    `yaml:"OpticalPowerTxLowWarningThreshold"`
-	OpticalPowerRxHighWarningThreshold int    `yaml:"OpticalPowerRxHighWarningThreshold"`
-	OpticalPowerRxLowWarningThreshold  int    `yaml:"OpticalPowerRxLowWarningThreshold"`
-	ProgOutputPowerMin                 int    `yaml:"ProgOutputPowerMin"`
-	ProgOutputPowerMax                 int    `yaml:"ProgOutputPowerMax"`
-	GridSpacingTxx                     int    `yaml:"GridSpacingTxx"`
-	CurrentLaserFrequencyTxx           int    `yaml:"CurrentLaserFrequencyTxx"`
-	TargetOutputPowerTxx               int    `yaml:"TargetOutputPowerTxx"`
+	SFF8024Identifier                  int     `yaml:"SFF8024Identifier"`
+	CmisRevision                       int     `yaml:"CmisRevision"`
+	ModuleRevision                     int     `yaml:"ModuleRevision"`
+	MediaType                          int     `yaml:"MediaType"`
+	VendorName                         string  `yaml:"VendorName"`
+	DateCode                           string  `yaml:"DateCode"`
+	MaxPower                           int     `yaml:"MaxPower"`
+	LenghtSMF                          int     `yaml:"LenghtSMF"`
+	ModuleTempMax                      float64 `yaml:"ModuleTempMax"`
+	ModuleTempMin                      float64 `yaml:"ModuleTempMin"`
+	TempMonHighWarningThreshold        float64 `yaml:"TempMonHighWarningThreshold"`
+	TempMonLowWarningThreshold         float64 `yaml:"TempMonLowWarningThreshold"`
+	VccMonHighWarningThreshold         float64 `yaml:"VccMonHighWarningThreshold"`
+	VccMonLowWarningThreshold          float64 `yaml:"VccMonLowWarningThreshold"`
+	OpticalPowerTxHighWarningThreshold float64 `yaml:"OpticalPowerTxHighWarningThreshold"`
+	OpticalPowerTxLowWarningThreshold  float64 `yaml:"OpticalPowerTxLowWarningThreshold"`
+	OpticalPowerRxHighWarningThreshold float64 `yaml:"OpticalPowerRxHighWarningThreshold"`
+	OpticalPowerRxLowWarningThreshold  float64 `yaml:"OpticalPowerRxLowWarningThreshold"`
+	ProgOutputPowerMin                 float64 `yaml:"ProgOutputPowerMin"`
+	ProgOutputPowerMax                 float64 `yaml:"ProgOutputPowerMax"`
+	GridSpacingTxx                     int     `yaml:"GridSpacingTxx"`
+	CurrentLaserFrequencyTxx           int     `yaml:"CurrentLaserFrequencyTxx"`
+	TargetOutputPowerTxx               float64 `yaml:"TargetOutputPowerTxx"`
 }
 
 type ModulesConfig struct {
@@ -38,8 +38,8 @@ type ModulesConfig struct {
 }
 
 type Step struct {
-	Endval   int `yaml:"endval"`
-	Duration int `yaml:"duration"`
+	Endval   float64 `yaml:"endval"`
+	Duration int     `yaml:"duration"`
 }
 
 type ScenarioModule struct {
